+++ source/apex_geometry.py
@@ -218,7 +218,7 @@ def apply_panel_deformation(panels, coeffs):
     return surface, normal, ds, rms_deform
 
 
-def generate_start_coeffs(random_key, panel_names, start_rms=1e-5):
+def generate_start_coeffs(random_key, panel_names, start_rms=1e-5, dtype=jnp.float32):
     coeffs_out = dict()
     for name in panel_names:
         random_key, subkey = jax.random.split(random_key)
@@ -226,7 +226,7 @@ def generate_start_coeffs(random_key, panel_names, start_rms=1e-5):
         coeffs = jax.random.uniform(subkey, shape=5, 
                                     minval=-start_rms, 
                                     maxval=start_rms)
-        coeffs_out[name] = coeffs
+        coeffs_out[name] = coeffs.astype(jnp.float32)
     return coeffs_out
